public_backend/requirements.txt
Flask==3.1.0
PyMySQL==1.1.1
python-dotenv==1.0.1
Flask-CORS==4.0.0
flasgger==0.9.7.1
PyJWT==2.10.1
